fixes #197 - moving tests.common to PyWPS distribution
